Revert "Admin Homepage: Dashboard Data Visualizations"
