+++ templates/accounts/totals-count.html
@@ -48,123 +48,9 @@
             </div>           
         </div>
 
-    </div>
-    
-    <div class="content-card-v2 content-card-space">
-        <h2>Accounts on the Hub</h2>
-        <hr class="searchbar-divider">
-
-        <div class="flex-this">
-            <div class="margin-right-16">
-                <form class="default-searchbar" method="GET" action="">
-                    <input type="search" name="q" placeholder="Search" autocomplete="off">
-                    <button class="search-btn"><i class="fa fa-search fa-lg" aria-hidden="true"></i></button>
-                </form>
-            </div>
-            <!-- <div class="flex-this align-center">
-                <form method="GET" action="" class="flex-this">
-                    <select name="sort" class="project-sort">
-                        <option
-                            class="darkteal-text pointer" 
-                            value="all"
-                        >
-                            Sort: All
-                        </option>
-                        <option
-                            class="darkteal-text pointer" 
-                            value="has_labels"
-                            {% if bool_dict.has_labels %} selected {% endif %}
-                        >
-                            Sort: Labels
-                        </option>
-                    </select>  
-                    <input class="primary-btn action-btn margin-left-8" type="submit" value="Sort"> 
-                </form>
-            </div> -->
-        </div>
-        <br>
-        <table class="projects-table">
-            <tr>
-                <th>Name</th>
-                <th>Location</th>
-                <th>Notices/Labels</th>
-                <th># of Projects</th>
-                <th># of Members</th>
-            </tr>
-            <tr>
-                <td>Column 1</td>
-                <td>Column 2</td>
-                <td>Column 3</td>
-                <td>Column 4</td>
-                <td>Column 5</td>
-            </tr>
-        </table>
-    </div>
+    </div>    
 
     <div class="content-card-v2 content-card-space">
-        <h2>Recent Project Action Updates</h2>
-        <hr class="searchbar-divider">
-
-        <div class="flex-this">
-            <div class="flex-this align-center">
-                <form method="GET" action="" class="flex-this">
-                    <select name="sort" class="project-sort">
-                        <option
-                            class="darkteal-text pointer" 
-                            value="all"
-                        >
-                            Filter: All
-                        </option>
-                        <option
-                            class="darkteal-text pointer" 
-                            value="all"
-                        >
-                            Filter: Label(s) Added
-                        </option>
-                        <option
-                            class="darkteal-text pointer" 
-                            value="all"
-                        >
-                            Filter: Notice(s) Added
-                        </option>
-                        <option
-                            class="darkteal-text pointer" 
-                            value="all"
-                        >
-                            Filter: Project Created
-                        </option>
-                    </select>  
-                    <input class="primary-btn action-btn margin-left-8" type="submit" value="Filter"> 
-                </form>
-                <p class="flex-this right-text">Total:100</p>
-            </div>
-        </div>
-        <br>
-        <table class="projects-table">
-            <tr>
-                <th>Action</th>
-                <th>Date/Time</th>
-            </tr>
-            <tr>
-                <td>TK A Label added to Project Name by Account Name</td>
-                <td>April 20, 2023 4:00PM EST</td>
-            </tr>
-            <tr>
-                <td>OTC Notice added to Account Name</td>
-                <td>April 20, 2023 4:00PM EST</td>
-            </tr>
-            <tr>
-                <td>Project Name created by Account Name</td>
-                <td>April 20, 2023 4:00PM EST</td>
-            </tr>
-            <tr>
-                <td>AI Notice added to Project Name by Account Name</td>
-                <td>April 20, 2023 4:00PM EST</td>
-            </tr>
-        </table>
-    </div>
-
-    <!-- <div class="content-card-v2 content-card-space">
         <h3>Open to Collaborate Notice URLs (Total: {{ otc_notices.count }})</h3>
         <div>
             {% if otc_notices %}
@@ -187,42 +73,6 @@
                     {% endfor %}
                 </ol>
             {% endif %}
-        </div>
-    </div> -->
-
-    <div class="content-card-v2 content-card-space">
-        <h2>Map</h2>
-        <hr class="searchbar-divider">
-        <div class="flex-this align-center">
-            <form method="GET" action="" class="flex-this">
-                <select name="sort" class="project-sort">
-                    <option
-                        class="darkteal-text pointer" 
-                        value="all"
-                    >
-                        Filter: All
-                    </option>
-                    <option
-                        class="darkteal-text pointer" 
-                        value="all"
-                    >
-                        Filter: Communities
-                    </option>
-                    <option
-                        class="darkteal-text pointer" 
-                        value="all"
-                    >
-                        Filter: Institutions
-                    </option>
-                    <option
-                        class="darkteal-text pointer" 
-                        value="all"
-                    >
-                        Filter: Researchers
-                    </option>
-                </select>  
-                <input class="primary-btn action-btn margin-left-8" type="submit" value="Filter"> 
-            </form>
         </div>
     </div>
