--- templates/admin/base_site.html
@@ -1,5 +1,0 @@
-{% extends 'admin/base_site.html' %} {% load static %}
-
-{% block extrastyle %}
-    <link rel="stylesheet" type='text/css' href="{% static 'css/admin.css' %}">
-{% endblock %}
