--- templates/admin/dashboard.html
@@ -1,42 +1,0 @@
-{% load static %}
-
-<script src="{% static 'javascript/admin.js' %}"></script>
-
-<div class="dashboard">
-    <div class="w-100">
-        <h2>Today: {% now "M. d, Y, H:i e" %}</h2>
-    </div>
-
-    <!-- Data Charts and Counts (Visualizations) -->
-    {% include 'admin/partials/_data-visuals.html' %}
-
-    <!-- Action Tables -->
-    {% include 'admin/partials/_action-tables.html' %}
-
-    <!-- ADMIN ACTIVITY LOG -->
-    <!-- <div class="module" id="recent-actions-module">
-        {% load log %}
-        {% get_admin_log 10 as admin_log for_user user %}
-        {% if not admin_log %}
-        <p>None Available</p>
-        {% else %}
-        <ul class="actionlist">
-        {% for entry in admin_log %}
-        <li class="{% if entry.is_addition %}addlink{% endif %}{% if entry.is_change %}changelink{% endif %}{% if entry.is_deletion %}deletelink{% endif %}">
-            {% if entry.is_deletion or not entry.get_admin_url %}
-                {{ entry.object_repr }}
-            {% else %}
-                <a href="{{ entry.get_admin_url }}">{{ entry.object_repr }}</a>
-            {% endif %}
-            <br>
-            {% if entry.content_type %}
-                <span class="mini quiet">{% filter capfirst %}{{ entry.content_type.name }}{% endfilter %}</span>
-            {% else %}
-                <span class="mini quiet">Unknown Content</span>
-            {% endif %}
-        </li>
-        {% endfor %}
-        </ul>
-        {% endif %}
-    </div> -->
-</div>
