--- templates/admin/nav_sidebar.html
@@ -1,5 +1,0 @@
-{% load i18n %}
-<nav class="sticky" id="nav-sidebar">
-  {% include 'admin/app_list.html' with app_list=available_apps show_changelinks=False %}
-</nav>
-<button class="sticky toggle-nav-sidebar" id="toggle-nav-sidebar" aria-label="{% translate 'Toggle navigation' %}"></button>
